Add files via upload
So i got the database on my home computer up and running 100% but the actual could we have doesn't have a main method. I went back to a previous version of it and i cant seem to find one that does. If someone could help out with that I can finish connecting the database.