Someone look at this please...
Finally got the login page hooked up and working. If someone could help
me to get the login page to go to the next page correctly, that would be
awesome cause i have to go to work and wont have too much more time to
work on it prior to our meeting.