Got it at least working...
we have a lot more work todo if we want to have the program all done
besides the algorithm. I have changed what we are mainly going to be
using for the ID to the email until we can find a way to actaully get
the ID into something that isnt an object.